requirements.txt
openai>=1.0.0
agents>=0.13.0
slack-sdk>=3.20.0
slack-bolt>=1.18.0
mcp-client>=0.4.0
pydantic>=2.0.0
asyncio
aiohttp>=3.8.0
python-dotenv>=1.0.0
jinja2>=3.1.0
markdown>=3.4.0
reportlab>=4.0.0
python-multipart>=0.0.6
fastapi>=0.100.0
uvicorn>=0.23.0 